feat!: basic impl of extendable thing
This makes Thing, its builder and many of the related structures generic
over "extension types". In a very simple case, it is possible to use a
custom struct for the `Other` generic in order to add custom properties
on the different layers. In general, it is possible to use an
heterogeneous list of _extension types_ in order to extend the `Thing`
multiple times -- notice that this is the default behavior when using
the helper functions, and the user never sees the eterogeneous list.

Currently, `ext` and `ext_with` methods have a temporary name, they
could be changed in the future.